apps/backend/pyproject.toml
[tool.poetry]
name = "backend-api"
version = "0.1.0"
description = ""
authors = ["YOOJIIN <yoojiin0328@naver.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11, <3.14"

fastapi = ">=0.117.1,<0.118.0"
uvicorn = ">=0.37.0,<0.38.0"

python-dotenv = ">=1.1.1,<2.0.0"
pandas = "^2.2.2"
sentence-transformers = "^3.0.1"
google-generativeai = "^0.7.1"
tqdm = "^4.66.4"

# --- 아래 라이브러리들은 현재 Gemini 스크립트에서 사용되지 않으며, ---
# --- 의존성 충돌의 직접적인 원인이 되어 잠시 주석 처리합니다. ---
# kiwipiepy = ">=0.21.0,<0.22.0"
# keybert = ">=0.9.0,<0.10.0"
# bertopic = ">=0.17.3,<0.18.0"
sqlalchemy = "^2.0.44"
psycopg = {extras = ["binary"], version = "^3.2.12"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
